+++ examples/example-mod/src/main/java/fr/d4emon/fenix/example/content/ModModels.java
@@ -16,6 +16,7 @@ public final class ModModels extends EmberModelProvider {
         cubeAll(ModBlocks.RUBY_BLOCK);
         cubeAll(ModBlocks.GLOWING_RUBY_BLOCK);
         cubeAll(ModBlocks.RUBY_TALLY);
+        cubeAll(ModBlocks.RUBY_SAFE);
 
         flatItem(ModItems.RUBY);
         flatItem(ModItems.RUBY_HAMMER);